Format dates in french
