+++ lib/emailer.js
@@ -1,6 +1,7 @@
 const nodemailer = require('nodemailer')
-const config = require('../config')
 
+const config = require('../config')
+const format = require('./format')
 
 const mailOptions = {
   debug: true,
@@ -37,17 +38,6 @@ const sendMail = async function (fromEmail, toEmail, subject, html) {
   });
 }
 
-// OVH uses format '00339112233', we want '09 11 22 33' for humans
-const formatFrenchPhoneNumber = machineReadableNumber => {
-  const lastNineDigits = machineReadableNumber.slice(-9)
-  const withPrefix = '0' + lastNineDigits
-  const splitByPairs = withPrefix.substring(0, 2)
-      + ' ' + withPrefix.substring(2, 4)
-      + ' ' + withPrefix.substring(4, 6)
-      + ' ' + withPrefix.substring(6, 8)
-  return splitByPairs
-}
-
 module.exports.sendConfCreatedEmail = async function(toEmail, confPhoneNumber, pin, freeAt) {
   const html = `
   <p>Bonjour,</p>
@@ -55,11 +45,11 @@ module.exports.sendConfCreatedEmail = async function(toEmail, confPhoneNumber, p
   <p>Votre conférence est bien créée.</p>
   <p>Pour vous y connecter : <p>
   <ul>
-    <li>appelez le <strong>${formatFrenchPhoneNumber(confPhoneNumber)}</strong> sur votre téléphone pro ou perso</li>
+    <li>appelez le <strong>${format.formatFrenchPhoneNumber(confPhoneNumber)}</strong> sur votre téléphone pro ou perso</li>
     <li>tapez le numéro de conférence à ${config.NUM_PIN_DIGITS} chiffres : <strong>${pin}</strong></li>
   </ul>
   <p></p>
-  <p>Cette conférence est active pendant ${config.CONFERENCE_DURATION_IN_MINUTES} minutes. Elle expirera à ${freeAt}.</p>
+  <p>Cette conférence est active pendant ${config.CONFERENCE_DURATION_IN_MINUTES} minutes. Elle expirera ${format.formatFrenchDate(freeAt)}.</p>
   <p>Elle peut accueillir jusqu'à 50 personnes.</p>
   <p></p>
   <p>Bonne journée avec ${config.APP_NAME} !</p>`
